refactor: extract auth lifecycle into AppDelegate+AuthLifecycle.swift, fix teardown gaps
